+++ src/classes/util/FormatUtils.ts
@@ -1,0 +1,70 @@
+/* Eurofurence Telecommunication Device Management System (EF-TDMS)
+ * Copyright (C) 2024 Niels Gandraß <niels@gandrass.de>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+/**
+ * Utility class for formatting data.
+ */
+export class FormatUtils {
+
+    /**
+     * Formats a MAC address to the common format.
+     * @param mac
+     */
+    public static formatMacAddress(mac: string): string {
+        if (mac.length === 12) {
+            return mac.replace(/(.{2})/g, '$1:').slice(0, -1).toUpperCase();
+        } else {
+            return mac.toUpperCase()
+        }
+    }
+
+    /**
+     * Format bytes as human-readable text.
+     *
+     * @param bytes Number of bytes.
+     * @param si True to use metric (SI) units, aka powers of 1000. False to use
+     *           binary (IEC), aka powers of 1024.
+     * @param dp Number of decimal places to display.
+     *
+     * @author mpen (https://stackoverflow.com/users/65387/mpen)
+     * @see https://stackoverflow.com/a/14919494
+     *
+     * @return Formatted string.
+     */
+    public static bytesToHumanReadableString(bytes: number, si: boolean = false, dp: number = 1) {
+        const thresh = si ? 1000 : 1024;
+
+        if (Math.abs(bytes) < thresh) {
+            return bytes + ' B';
+        }
+
+        const units = si
+            ? ['kB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB']
+            : ['KiB', 'MiB', 'GiB', 'TiB', 'PiB', 'EiB', 'ZiB', 'YiB'];
+        let u = -1;
+        const r = 10 ** dp;
+
+        do {
+            bytes /= thresh;
+            ++u;
+        } while (Math.round(Math.abs(bytes) * r) / r >= thresh && u < units.length - 1);
+
+
+        return bytes.toFixed(dp) + ' ' + units[u];
+    }
+
+}